Some event based additions

+++ include/dolphin/interpreter/processor.hpp
@@ -55,7 +55,7 @@ namespace Toolbox::Interpreter {
     }
 
     inline bool MemoryContainsPAddress(const Buffer &buffer, s32 address) {
-        return address >= 0 && address < buffer.size();
+        return address >= 0 && address < (s32)buffer.size();
     }
 
     inline bool IsRegValid(u8 reg) { return reg < 32; }
